requirements.txt
Django==5.0.2
djangorestframework==3.14.0
schedule==1.2.1
gunicorn==21.2.0
requests==2.31.0
discord.py==2.3.2
